chore: update CourseExtensionIdeaModel in client/cmd/login.go
- Remove unused imports and properties
- Update type declarations and imports
- Update getter methods for development status
- Update isNewlyCreated method to use a more readable condition
- Update vote method to increment votesCount and create a new record
- Update unvote method to decrement votesCount and unload records
